EVL-211: extract the shared SurveyWidget shell
The onboarding "How did you hear about Hoop?" survey and the TTFV survey are
the same widget: a launcher pinned to the right edge of the viewport that
expands into a non-blocking card, with an acknowledgement that retires it.
Pull that out of OriginSurvey so TTFV can be a second consumer rather than a
copy.

SurveyWidget owns the presentation and the collapsed/open/closed lifecycle.
It does not own visibility or submission — the survey feature passes `due`
(the gateway's verdict) and flips `answered` once the write is on its way.
SurveyWidget.ChoiceQuestion carries the part worth sharing: a single choice
with a free-text follow-up on one option, and the interlock that keeps Submit
disabled until that text is filled in.

ChoiceQuestion is controlled rather than self-contained. The shell unmounts
the card when the user collapses it with the X, so state kept inside would be
wiped by a mis-click — which is exactly what the original set out to avoid.
Holding it in the feature, which stays mounted, is what makes collapsing free.

OriginSurvey is unchanged in behaviour; it is the same JSX reached through the
shell.

One real fix along the way: the card's max-height referenced
--mantine-spacing-xxl, which does not exist. theme.js adds the 4/8/16/24/32/
48/64 scale under *Alt keys and leaves Mantine's own xs-xl defaults in place,
so the variable was undefined — which invalidates the whole calc() and drops
the declaration. The card had no max-height at all and ran off the top of
short viewports instead of scrolling. The CLAUDE.md spacing table is what led
to it, so it is corrected here too, with a note on why a misspelled variable
fails silently.
